* Import directory corruptions are now checked and reported * Resource corruptions are now checked and reported * MS Visual Studio can now build the library without warning. Phew.

+++ include/pelib/InputBuffer.h
@@ -39,7 +39,7 @@ namespace PeLib
 
 			std::vector<char> data(sizeof(T), 0);
 			std::size_t size = ulIndex + sizeof(T) > m_vBuffer.size() ? m_vBuffer.size() - ulIndex : sizeof(T);
-			read(data.data(), size);
+			read(data.data(), (unsigned long)size);
 			value = *reinterpret_cast<T*>(data.data());
 			return *this;
 		  }
